last

+++ src/main/java/ru/chat/service/UserService.java
@@ -2,14 +2,16 @@ package ru.chat.service;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Component;
 import ru.chat.dto.userDTO.RegUserDTO;
-import ru.chat.dto.userDTO.UpdateUserDTO;
+import ru.chat.dto.userDTO.UserUpdateDTO;
 import ru.chat.entity.User;
+import ru.chat.entity.enums.AppRole;
 import ru.chat.mapper.UserMapper;
+import ru.chat.repository.UserInChatRepository;
 import ru.chat.repository.UserRepository;
+import ru.chat.service.exception.YouDontHavePermissionExceptiom;
 
 import javax.transaction.Transactional;
 import java.security.Principal;
@@ -23,6 +25,12 @@ public class UserService {
 
     private final UserMapper userMapper;
     private final UserRepository userRepository;
+    private final UserInChatRepository userInChatRepository;
+
+    private User fromPrincipal(Principal principal) {
+        return userRepository.findByEmail(principal.getName())
+                .orElseThrow(() -> new UsernameNotFoundException(""));
+    }
 
     public User create(RegUserDTO userDTO) {
         User user = userRepository.save(userMapper.create(userDTO));
@@ -36,24 +44,35 @@ public class UserService {
         log.info("Get user - {} with id - {}", user, id);
         return user;
     }
+
     public List<User> getAll() {
         log.info("Get all users");
         return userRepository.findAll();
     }
 
-    public void delete(Long id) {
-        User user = userRepository.getById(id);
-        userRepository.delete(user);
-        log.info("User with id - {} was deleted", id);
+    // * удаление других пользователей для админов приложения
+    public void delete(Long id, Principal principal) throws YouDontHavePermissionExceptiom {
+        User admin = this.fromPrincipal(principal);
+
+        if (admin.getRole() == AppRole.ROLE_ADMIN) {
+            User user = userRepository.getById(id);
+            userRepository.delete(user);
+            userInChatRepository.deleteAllByUser(user);
+            log.info("User with id - {} was deleted", id);
+        } else {
+            throw new YouDontHavePermissionExceptiom("Only admin can delete other user from app");
+        }
     }
 
+    // * удаление себя
     public void delete(Principal principal) {
         User user = this.fromPrincipal(principal);
         userRepository.delete(user);
-        log.info("Current user was deleted (id - {})", user.getId());
+
+        log.info("Current user was deleted (username - {})", user.getUsername());
     }
 
-    public UpdateUserDTO update(UpdateUserDTO userDTO, Principal principal) {
+    public UserUpdateDTO update(UserUpdateDTO userDTO, Principal principal) {
         User user = fromPrincipal(principal);
 
         user.setEmail(userDTO.getEmail());
@@ -65,8 +84,10 @@ public class UserService {
         return userDTO;
     }
 
-    public User fromPrincipal(Principal principal) {
-        return userRepository.findByEmail(principal.getName())
-                .orElseThrow(() -> new UsernameNotFoundException(""));
+    public Object getCurrent(Principal principal) {
+        User user = this.fromPrincipal(principal);
+
+        log.info("{} go into profile", user.getUsername());
+        return userMapper.toUserResponseDTO(user);
     }
 }
